+++ common/src/main/java/io/github/jamalam360/reaping/ReapingExpectPlatform.java
@@ -4,6 +4,7 @@ import dev.architectury.injectables.annotations.ExpectPlatform;
 import io.github.jamalam360.reaping.config.ReapingConfig;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.damage.DamageSource;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.loot.context.LootContext;
 
 public class ReapingExpectPlatform {
@@ -21,4 +22,6 @@ public class ReapingExpectPlatform {
     public static  LootContext.Builder getLootContextBuilder(LivingEntity entity, boolean causedByPlayer, DamageSource source) {
         throw new AssertionError();
     }
+
+    public void setScale(PlayerEntity player, double scale) { throw new AssertionError(); }
 }